refactor(opal): migrate time utilities to @opal/time
Adds a new `@opal/time` entry point to the Opal library containing all
pure time-formatting and duration utilities previously in `web/src/lib/time.ts`.
`getSecondsUntilExpiration` moves to `web/src/lib/user.ts` as it is
user/auth-specific. All 27 callsites updated accordingly.
