+++ web/src/app/admin/scim/ScimSyncCard.tsx
@@ -4,7 +4,7 @@ import { Section } from "@/layouts/general-layouts";
 import Card from "@/refresh-components/cards/Card";
 import { Button, Divider } from "@opal/components";
 import Text from "@/refresh-components/texts/Text";
-import { timeAgo } from "@/lib/time";
+import { timeAgo } from "@opal/time";
 
 // ---------------------------------------------------------------------------
 // Props
