+++ web/src/app/ee/admin/performance/usage/UsageReports.tsx
@@ -23,7 +23,7 @@ import React, { useState } from "react";
 import { UsageReport } from "./types";
 import { ThreeDotsLoader } from "@/components/Loading";
 import Link from "next/link";
-import { humanReadableFormat, humanReadableFormatWithTime } from "@/lib/time";
+import { humanReadableFormat, humanReadableFormatWithTime } from "@opal/time";
 import { ErrorCallout } from "@/components/ErrorCallout";
 import { PageSelector } from "@/components/PageSelector";
 import { Divider } from "@opal/components";
