+++ web/src/refresh-components/buttons/source-tag/SourceTagDetailsCard.tsx
@@ -12,7 +12,7 @@ import {
 import { SourceIcon } from "@/components/SourceIcon";
 import { WebResultIcon } from "@/components/WebResultIcon";
 import { ValidSources } from "@/lib/types";
-import { timeAgo } from "@/lib/time";
+import { timeAgo } from "@opal/time";
 import type { IconProps } from "@opal/types";
 import { SubQuestionDetail } from "@/app/app/interfaces";
